Move class so we can reduce vis and remove checkstyle javadoc requirement
